course detail page created

+++ lib/common/widgets/base_text_widget.dart
@@ -4,7 +4,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import '../values/colors.dart';
 
 Widget reusableText(String text ,
-    {Color color =AppColors.primaryText,int fontsize = 16,FontWeight fontWeight = FontWeight.bold}){
+    {Color color =AppColors.primaryText,double fontsize = 16,FontWeight fontWeight = FontWeight.bold}){
   return Text(
     text,
     style: TextStyle(
